Searchable Knowledge Graph/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
spacy==3.7.2
python-dotenv==1.0.1
neo4j==5.13.0
PyMuPDF==1.23.7
beautifulsoup4==4.12.2
requests==2.31.0
